ath10k: Fix mgmt tx status for 10.4.3 firmware.
When testing a 10.4.3 firmware in station mode, I notice that
when the AP is powered down, the ath10k does not notice AP is gone
because mgt frames get tx status of 3, which is not handled.

It appears that status 3 means something similar to failed-retry.

Treating it thus lets the station disconnect properly.

Tested against a non-stock 10.4.3 firmware, but likely upstream
firmware acts similarly in this case.

Signed-off-by: Ben Greear <greearb@candelatech.com>